+++ Runtime/World/Components/Renderable.cpp
@@ -90,7 +90,7 @@ namespace Spartan
         );
     }
 
-    Renderable::Renderable(Context* context, Entity* entity, uint32_t id /*= 0*/) : IComponent(context, entity, id)
+    Renderable::Renderable(Context* context, Entity* entity, uint64_t id /*= 0*/) : IComponent(context, entity, id)
     {
         m_geometry_type         = Geometry_Custom;
         m_geometryIndexOffset   = 0;
