+++ Runtime/World/Components/Terrain.cpp
@@ -40,7 +40,7 @@ using namespace Spartan::Math;
 
 namespace Spartan
 {
-    Terrain::Terrain(Context* context, Entity* entity, uint32_t id /*= 0*/) : IComponent(context, entity, id)
+    Terrain::Terrain(Context* context, Entity* entity, uint64_t id /*= 0*/) : IComponent(context, entity, id)
     {
 
     }
