+++ Runtime/World/Components/Terrain.h
@@ -38,7 +38,7 @@ namespace Spartan
     class SPARTAN_CLASS Terrain : public IComponent
     {
     public:
-        Terrain(Context* context, Entity* entity, uint32_t id = 0);
+        Terrain(Context* context, Entity* entity, uint64_t id = 0);
         ~Terrain() = default;
 
         //= IComponent ===============================
